+++ src/taurenmd/libs/libcalc.py
@@ -26,7 +26,6 @@ import math
 import numpy as np
 from MDAnalysis.analysis.rms import RMSD as mdaRMSD
 from MDAnalysis.analysis.rms import RMSF as mdaRMSF
-from pyquaternion import Quaternion as Q
 
 from taurenmd import core as tcore
 from taurenmd import log
@@ -252,105 +251,9 @@ def calc_planes_angle(a1, b1, c1, a2, b2, c2, aunit='radians'):
         return angle
 
 
-#@libcli.add_reference(tcore.ref_pyquaternion)
-#def generate_quaternion_rotations(
-#        rotation_axis,
-#        rotating_vector,
-#        start=0,
-#        end=360,
-#        num=360 * 3,
-#        endpoint=True,
-#        ):
-#    """
-#    Generate quaternion rotations of a vector around and axis.
-#
-#    Rotates a `vector` around an `axis` for a series of angles.
-#    Rotation is performed using `Quaternion rotation <https://en.wikipedia.org/wiki/Quaternions_and_spatial_rotation>`_;
-#    namely, `pyquaterion.rotate <http://kieranwynn.github.io/pyquaternion/#rotation>`_.
-#
-#    If you use this function you should cite
-#    `PyQuaternion package <http://kieranwynn.github.io/pyquaternion/>`_.
-#
-#    Parameters
-#    ----------
-#    rotation_axis : tuple, list or array-like
-#        The XYZ coordinates of the rotation axis. This is the axis
-#        around which the vector will be rotated.
-#
-#    rotation_vector : tuple, list or array-like
-#        The XYZ coordinates of the vector to rotate.
-#
-#    start : int
-#        The starting rotation angle in degrees in the series.
-#        Defaults to ``0``.
-#
-#    end : int
-#        The final rotation angle in degrees in the series.
-#        Defaults to ``360``.
-#
-#    num : int
-#        The number of rotation steps in the angle rotation series.
-#        This is provided by ``np.linspace(start, end, num=num)``.
-#
-#    Returns
-#    -------
-#    list of 2 unit tuples
-#        Where the list indexes follow the progression along
-#        ``np.linspace(start, end, num=num)`` and tuple elements
-#        are the rotation quatertion used to rotate ``rotation_vector``
-#        and the resulting rotated vector in its unitary form.
-#    """  # noqa: E501
-#    vec_u = Q(vector=rotating_vector).unit
-#    rot_u = Q(vector=rotation_axis).unit
-#
-#    Q_rotated_tuples = []
-#    for angle in np.linspace(start, end, num=num, endpoint=endpoint):
-#        qq = Q(axis=rot_u.vector, degrees=angle)
-#        Q_rotated_tuples.append((qq, qq.rotate(vec_u)))
-#
-#    return Q_rotated_tuples
-#
-#
-#@libcli.add_reference(tcore.ref_pyquaternion)
-#def sort_by_minimum_Qdistances(rotation_tuples, reference_vector):
-#    """
-#    Sort a list of quaternion rotation tuples.
-#
-#    By its distance to a ``reference_vector``. Designed to receive
-#    the output of :py:func:`generate_quaternion_rotations`.
-#
-#    If you use this function you should cite
-#    `PyQuaternion package <http://kieranwynn.github.io/pyquaternion/>`_.
-#
-#    Parameters
-#    ----------
-#    rotation_tuples : list of tuples
-#        List of tuples containing Rotation quaternions and resulting
-#        rotated vector. In other words, the output from
-#        :py:func:`generate_quaternion_rotations`.
-#
-#    reference_vector : tuple, list or array-like
-#        The XYZ 3D coordinates of the reference vector. The quaternion
-#        distance between vectors in ``rotation_tuples`` and this
-#        vector will be computed.
-#
-#    Returns
-#    -------
-#    list
-#        The list sorted according to the creterion.
-#    """
-#    ref_u = Q(vector=reference_vector).unit
-#
-#    minimum = sorted(
-#        rotation_tuples,
-#        key=lambda x: math.degrees(Q.distance(x[1], ref_u)) % 360,
-#        )
-#    return minimum
-
-
 def calc_torsion_angles(coords):
     """
-    Calculates torsion angles from sequential coordinates.
+    Calculate torsion angles from sequential coordinates.
 
     Uses ``NumPy`` to compute angles in a vectorized fashion.
     Sign of the torsion angle is also calculated.