+++ tests/test_clis.py
@@ -495,15 +495,9 @@ def test_cli_rotations_1():
         export='rots.csv',
         )
 
-    p1 = Path('roll_angles_rots.csv')
-    p2 = Path('pitch_angles_rots.csv')
-    p3 = Path('yaw_angles_rots.csv')
+    p1 = Path('rots.csv')
     assert p1.exists()
-    assert p2.exists()
-    assert p3.exists()
     p1.unlink()
-    p2.unlink()
-    p3.unlink()
 
 
 def test_cli_rotations_2():