+++ tests/test_libcalc.py
@@ -2,7 +2,6 @@
 import MDAnalysis as mda
 import numpy as np
 import pytest
-from pyquaternion import Quaternion as Q
 
 from taurenmd.libs import libcalc as lc
 
@@ -96,45 +95,3 @@ def test_calc_planes_angle(a1, b1, c1, a2, b2, c2, aunit, expected):
     """Test angle between planes."""
     result = lc.calc_planes_angle(a1, b1, c1, a2, b2, c2, aunit)
     assert abs(result - expected) < 0.00001
-
-
-def test_gen_quaternion_rot():
-    """Test quaternion rotation."""
-    result = lc.generate_quaternion_rotations(
-        np.array([1.0, 0.0, 0.0]),
-        np.array([0.0, 1.0, 0.0]),
-        num=360
-        )
-    assert isinstance(result, list)
-    assert len(result) == 360
-    assert isinstance(result[0], tuple)
-    assert len(result[0]) == 2
-    assert isinstance(result[0][0], Q)
-    assert isinstance(result[0][1], Q)
-
-
-def test_sort_by_minimum():
-    """Test sort quaternion rotations."""
-    rotations = lc.generate_quaternion_rotations(
-        np.array([1.0, 0.0, 0.0]),
-        np.array([0.0, 1.0, 0.0]),
-        num=10,
-        endpoint=False,
-        )
-    for i in range(len(rotations)):
-        print(rotations[i][0].degrees)
-    result = lc.sort_by_minimum_Qdistances(
-        rotations,
-        np.array([1.0, 0.0, 0.0]),
-        )
-
-    for i in range(len(result)):
-        print(result[i][0].degrees)
-
-    assert isinstance(result, list)
-    assert isinstance(result[0], tuple)
-    assert len(result[0]) == 2
-    assert isinstance(result[0][0], Q)
-    assert isinstance(result[0][1], Q)
-    assert result[0][0].degrees < 0.001
-    assert abs(result[0][-1].degrees - 180) < 0.001