+++ CHANGELOG.rst
@@ -10,6 +10,9 @@ Upon version 0.8, and before version 1, SV2 major version increments are reflect
 Changelog
 =========
 
+v0.9.7 (2021-11-20)
+------------------------------------------------------------
+
 * improved trajectory report (#54)
 
 v0.9.6 (2021-02-22)